Home

MonteCarloX.jl Documentation

MonteCarloX.jl Documentation

Equilibrium Tools

Non-equilibrium Tools

Helper Modules and Functioms

Equilibrium Tools

Importance Sampling (Metropolis)

MonteCarloX.Metropolis.accept — Method.

accept(log_weight::Function, args_new::Tuple{Number, N}, args_old::Tuple{Number, N}, rng::AbstractRNG)::Bool where N

Evaluate acceptance probability according to Metropolis criterium for imporance sampling of $P(E) \propto e^{log\_weight(E)}$.

Arguments

log_weight(args): logarithmic ensemble weight function, e.g., canomical ensemble $log\_weight(E) = -\beta E$

args_new: arguments (can be Number or Tuple) for new (proposed) state

args_old: arguments (can be Number or Tuple) for old state

rng: random number generator, e.g. MersenneTwister

source

MonteCarloX.Metropolis.random_element — Method.

random_element(list_probabilities::Vector{Float64},rng::AbstractRNG)::Int

Pick an index from a list of probabilities.

source

MonteCarloX.Metropolis.sweep — Method.

sweep(list_updates, list_probabilities::Vector{AbstractFloat}, rng::AbstractRNG; number_updates::Int=1)

Randomly pick und run update (has to check acceptance by itself!) from list_updates with probability specified in list_probabilities and repeat this number_updates times.

source

Reweighting

MonteCarloX.Reweighting — Module.

MonteCarloX.Reweighting

Module for timeseries and histogram reweighting

source

MonteCarloX.Reweighting.binary_search — Method.

binary_search(sorted::AbstractVector{T}, value::T)::Int where {T<:Real}

Perfom a binary search to return the index i of an sorted array such that sorted[i-1] < value <= sorted[i]

Examples

julia> MonteCarloX.binary_search([1.,2.,3.,4.],2.5)
3

julia> MonteCarloX.binary_search([1,2,3,4],2)
2

source

MonteCarloX.Reweighting.canonical_timeseries — Method.

timeseries reweighting in canonical ensemble

methods:

metropolis (for each temperature separate)

<O> = sum Oi Ptarget(Ei)/Psource(Ei) / sum Ptarget(Ei)/Psource(args_i)

source

MonteCarloX.Reweighting.distribution_from_timeseries_log — Method.

Estimate distribution from a list of (measured) arguments to an (n-dimensional) probability distribution

For higher dimensional distributions (e.g. P(E,M)) list_args needs to be a list of tuples

returns:

Dictionary

source

MonteCarloX.Reweighting.expectation_value_from_histogram_log — Method.

Estimate expectation value from histogram (

Ref: Janke

Todo: rename logP... to logweight!!

important: histobs(args) = sum Oi delta(args - args_i) hists are dictionaries? can this be generalized to higher dimensions? nd histograms as dictionary?

source

MonteCarloX.Reweighting.expectation_value_from_timeseries — Method.

timeseries reweighting

wrapper for timeseries rewighting with logarithmic distributions

#Arguments

source

MonteCarloX.Reweighting.expectation_value_from_timeseries_log — Method.

timeseries reweighting

can be used with methods:

metropolis (for each temperature separate)

parallel tempering (for each temperature separate)

multicanonical

population annealing?

canonical <O> = sum Oi Ptarget(Ei)/Psource(Ei) / sum Ptarget(Ei)/Psource(args_i)

or in general <O> = sum Oi Ptarget(argsi)/Psource(argsi) / sum Ptarget(argsi)/Psource(args_i)

source

MonteCarloX.Reweighting.log_sum — Method.

log_sum(a::T,b::T)

Return result of logarithmic sum $c = \ln(A+B) = a + \ln(1+e^{|b-a|})$ where $C = e^c = A+B = e^a + e^b$.

This is useful for sums that involve elements that span multiple orders of magnitude, e.g., the partition sum that is required as normalization factor during reweighting.

Examples

julia> exp(MonteCarloX.log_sum(log(2.), log(3.)))
5.000000000000001

source

Non-equilibrium Tools

Kintetic Monte-Carlo

MonteCarloX.KineticMonteCarlo — Module.

MonteCarloX.KineticMonteCarlo

Module that allows to implement (rejection-free) kinetic Monte Carlo algorithms in general

source

MonteCarloX.KineticMonteCarlo.binary_search — Method.

binary_search(sorted::AbstractVector{T}, value::T)::Int where {T<:Real}

Perfom a binary search to return the index i of an sorted array such that sorted[i-1] < value <= sorted[i]

Examples

julia> MonteCarloX.binary_search([1.,2.,3.,4.],2.5)
3

julia> MonteCarloX.binary_search([1,2,3,4],2)
2

source

MonteCarloX.KineticMonteCarlo.log_sum — Method.

log_sum(a::T,b::T)

Return result of logarithmic sum $c = \ln(A+B) = a + \ln(1+e^{|b-a|})$ where $C = e^c = A+B = e^a + e^b$.

This is useful for sums that involve elements that span multiple orders of magnitude, e.g., the partition sum that is required as normalization factor during reweighting.

Examples

julia> exp(MonteCarloX.log_sum(log(2.), log(3.)))
5.000000000000001

source

MonteCarloX.KineticMonteCarlo.next_event — Method.

generate events (dt,id) from a list of rates such that their occurence corresponds with their rate

source

MonteCarloX.KineticMonteCarlo.next_event — Method.

fast implementation of nexteventrate if sum(list_rates) is known

source

MonteCarloX.KineticMonteCarlo.next_event_id — Method.

next event id for list of stochastic events described by occurence rates and the sum over those

source

MonteCarloX.KineticMonteCarlo.next_event_id — Method.

next event id for list of stochastic events describes by cumulated rates vector

source

MonteCarloX.KineticMonteCarlo.next_event_time — Method.

next event time for Poisson process with given rate

Arguments

source

MonteCarloX.Gillespie — Module.

MonteCarloX.Gillespie

Module that allows to implement Gillespie algorithm for the statistical time trajectory of a stochastic equation Uses kinetic Monte Carlo

source

MonteCarloX.Gillespie.advance — Method.

Advance the system by time T

The function generates new events as long as the total time increment does not go beyond the total time T. If the next event would advance the time too far, then this event is NOT performed. The system is thus in a state that it would have after time T.

Assuming Poisson rates, the function can be called again to advance from time T because generation of events is independent. (TODO: check)

source

MonteCarloX.Gillespie.next_event_rate — Method.

fast implementation of nexteventrate if sum(list_rates) is known

source

MonteCarloX.Gillespie.next_event_rate — Method.

generate events (dt,id) from a list of rates such that their occurence corresponds with their rate

source

Inhomogeneous Poisson Process

MonteCarloX.InhomogeneousPoissonProcess — Module.

Module MonteCarloX.InhomogeneousPoissonProcess

basic functions to evaluate (a coupled set of) inhomogeneous Poisson processes with time-dependent rate function rate(t)

source

MonteCarloX.InhomogeneousPoissonProcess.next_event_time — Method.

nexteventtime(rate::Function, max_rate::Float64, rng::AbstractRNG)::Float64

Generate a new event from an inhomogeneous poisson process with rate Lambda(t). Based on (Ogata’s Modified Thinning Algorithm: Ogata, 1981, p.25, Algorithm 2) see also https://www.math.fsu.edu/~ychen/research/Thinning%20algorithm.pdf

Arguments

rate: rate(dt) has to be defined outside (e..g t-> rate(t+t0,args))

max_rate: maximal rate in near future (has to be evaluated externally)

rng: random number generator

API - output

returns the next event time

source

MonteCarloX.InhomogeneousPoissonProcess.next_event_time_for_piece_wise_decreasing_rate — Method.

Generate a new event from an inhomogeneous poisson process with rate Lambda(t) under the assumption that rate(dt) is monotonically decreasing. Based on (Ogata’s Modified Thinning Algorithm: Ogata, 1981, p.25, Algorithm 3) see also https://www.math.fsu.edu/~ychen/research/Thinning%20algorithm.pdf

Arguments

rate: rate(dt) has to be defined outside (e..g t-> rate(t+t0,args))

rng: random number generator

API - output

returns the next event time

source

Helper Modules and Functioms

Histograms

MonteCarloX.Histograms — Module.

MonteCarloX.Histograms

Module represents histograms/distributions as dictionaries and handles addition

source

MonteCarloX.Histograms.distribution — Method.

create a distribution (sum_args dist[args] = 1) from a list of values

source

MonteCarloX.Histograms.histogram — Method.

create a histogram from a list of values

can this be formulated for listargs, dargs, argsref

source

Remaining helper

MonteCarloX.binary_search — Method.

binary_search(sorted::AbstractVector{T}, value::T)::Int where {T<:Real}

Perfom a binary search to return the index i of an sorted array such that sorted[i-1] < value <= sorted[i]

Examples

julia> MonteCarloX.binary_search([1.,2.,3.,4.],2.5)
3

julia> MonteCarloX.binary_search([1,2,3,4],2)
2

source

MonteCarloX.log_sum — Method.

log_sum(a::T,b::T)

Return result of logarithmic sum $c = \ln(A+B) = a + \ln(1+e^{|b-a|})$ where $C = e^c = A+B = e^a + e^b$.

This is useful for sums that involve elements that span multiple orders of magnitude, e.g., the partition sum that is required as normalization factor during reweighting.

Examples

julia> exp(MonteCarloX.log_sum(log(2.), log(3.)))
5.000000000000001

source