Importance Sampling (Metropolis)

MonteCarloX.accept — Function

accept(log_weight::Function, args_new::Tuple{Number, N}, args_old::Tuple{Number, N}, rng::AbstractRNG)::Bool where N

Evaluate most general acceptance probability for imporance sampling of $P(x) \propto e^{\mathrm{log\_weight}(x)}$.

Arguments

log_weight(args): logarithmic ensemble weight function, e.g., canomical ensemble $\mathrm{log\_weight}(x) = -\beta x$

args_new: arguments (can be Number or Tuple) for new (proposed) state

args_old: arguments (can be Number or Tuple) for old state

rng: random number generator, e.g. MersenneTwister

Specializations

accept(alg::Metropolis(), rng::AbstractRNG, xnew::T, xold::T) where T (@ref)

accept(alg::Metropolis, rng::AbstractRNG, beta::Float64, dx::T)::Bool where T

Standard metropolis algorithm with $p(x\to x^\prime) = \mathrm{min} (1, e^{-\beta \Delta x} )$ where $\Delta x = x^\prime - x$

accept(alg::MetropolisHastings, rng::AbstractRNG, beta::Float64, dx::T, q_old::Float64, q_new::Float64)::Bool where T

MonteCarloX.sweep — Function

sweep(list_updates, list_weights::AbstractWeights, rng::AbstractRNG; number_updates::Int=1) where T<:AbstractFloat

Randomly pick und run update (has to check acceptance by itself!) from list_updates with probability specified in list_probabilities and repeat this number_updates times.

Remarks

Walker's alias method?

Reweighting

MonteCarloX.expectation_value_from_timeseries — Function

expectation_value_from_timeseries(log_P_target::Function, log_P_source::Function, list_args, list_obs::Vector{Tin})::Tout where {Tin<:Number,Tout<:AbstractFloat}

Calculate the expectation value of an observable in P_target from a list of measured observables in P_source.

This can be used for observables measured in equilibrium, e.g., from methods:

metropolis (for each temperature separate)

parallel tempering (for each temperature separate)

multicanonical

population annealing

Background

Definition of reweighting in general:

\[\langle O\rangle = \sum O_i P_\mathrm{target}(args_i)/P_\mathrm{source}(args_i) / \sum P_\mathrm{target}(args_i)/P_\mathrm{source}(args_i)\]

Definition of reweighting for the canonical ensemble:

\[\langle O\rangle = \sum O_i e^{\beta_\mathrm{target} E_i - \beta_\mathrm{source} E_i} / \sum e^{\beta_\mathrm{target} E_i - \beta_\mathrm{source} E_i}\]

Remark

So far, this may not be well implemented for type stability. However, it should not be the most timeconsuming part of the simulation so this problem is moved to later time.

MonteCarloX.distribution_from_timeseries — Function

distribution_from_timeseries(log_P_target, log_P_source, list_args, range [, mode])

estimate reweighted distribution over range for target weights log_P_target from a timeseries list_args that was measured with log_P_source. Intended also for n-dimensional probability distributions, e.g. P(E,M), where list_args needs to be a list of tuples

returns: - StatsBase.Histogram