Home

MonteCarloX.jl

MonteCarloX.binary_search

MonteCarloX.log_sum

MonteCarloX.binary_search — Method.

binary_search(cumulated_sum::AbstractVector{T}, value::T)::Int where {T<:Real}

Perfom a binary search to return the index i of an ordered array (cumulatedsum) such that cumulatedsum[i-1] < value <= cumulated_sum[i]

Examples

julia> MonteCarloX.binary_search([[1.,2.,3.,4.],2.5)
3

julia> MonteCarloX.binary_search([[1,2,3,4],2)
2

source

MonteCarloX.log_sum — Method.

log_sum(a::T,b::T)

Return result of logarithmic sum $c = ln(A+B) = a + ln(1+e^{|b-a|})$ where $C = e^c = A+B = e^a + e^b$.

This is useful for sums that involve elements that span multiple orders of magnitude, e.g., the partition sum that is required as normalization factor during reweighting.

Examples

julia> exp(MonteCarloX.log_sum(log(2.), log(3.)))
5.000000000000001

source