Helper

Helper Tools

MonteCarloX.binary_search

MonteCarloX.log_sum

MonteCarloX.random_element

Full Docs

MonteCarloX.binary_search — Method.

binary_search(sorted::AbstractVector{T}, value::T)::Int where {T<:Real}

Perfom a binary search to return the index i of an sorted array such that sorted[i-1] < value <= sorted[i]

Examples

julia> MonteCarloX.binary_search([1.,2.,3.,4.],2.5)
3

julia> MonteCarloX.binary_search([1,2,3,4],2)
2

source

MonteCarloX.log_sum — Method.

log_sum(a::T,b::T)

Return result of logarithmic sum $c = \ln(A+B) = a + \ln(1+e^{|b-a|})$ where $C = e^c = A+B = e^a + e^b$.

This is useful for sums that involve elements that span multiple orders of magnitude, e.g., the partition sum that is required as normalization factor during reweighting.

Examples

julia> exp(MonteCarloX.log_sum(log(2.), log(3.)))
5.000000000000001

source

MonteCarloX.random_element — Method.

random_element(list_probabilities::Vector{T},rng::AbstractRNG)::Int where T<:AbstractFloat

Pick an index with probability defined by list_probability (which needs to be normalized).

#Remark Deprecated for use of StatsBase.sample

Examples

julia> using Random

julia> rng = MersenneTwister(1000);

julia> MonteCarloX.random_element([0.1,0.2,0.3,0.4],rng)
4
julia> MonteCarloX.random_element([0.1,0.2,0.3,0.4],rng)
4
julia> MonteCarloX.random_element([0.1,0.2,0.3,0.4],rng)
3
julia> MonteCarloX.random_element([0.1,0.2,0.3,0.4],rng)
4

source