Kintetic Monte-Carlo

MonteCarloX.next — Function

next([rng::AbstractRNG,] alg::KineticMonteCarlo, rates::AbstractWeights)::Tuple{Float64,Int}

Next stochastic event ($\Delta t$, index) drawn proportional to probability given in rates

next([rng::AbstractRNG,] alg::KineticMonteCarlo, event_handler::AbstractEventHandlerRate)

Next stochastic event ($\Delta t$, event type) organized by event_handler fast(to be tested, depends on overhead of EventList) implementation of nexteventrate if defined by EventList object

Generate a new event from a collection of inhomogeneous poisson processes with rates Lambda(t).

Arguments

rates: rates(dt); Float -> [Float]

max_rate: maximal rate in near future (has to be evaluated externally)

rng: random number generator

API - output

returns the next event time and event id as tuple (dt, id)

MonteCarloX.next_time — Function

next_time([rng::AbstractRNG,] rate::Float64)::Float64

Next stochastic $\Delta t$ for Poisson process with rate

nexteventtime(rate::Function, max_rate::Float64, rng::AbstractRNG)::Float64

Generate a new event from an inhomogeneous poisson process with rate Lambda(t). Based on (Ogata’s Modified Thinning Algorithm: Ogata, 1981, p.25, Algorithm 2) see also https://www.math.fsu.edu/~ychen/research/Thinning%20algorithm.pdf

Arguments

rate: rate(dt) has to be defined outside (e..g t-> rate(t+t0,args))

max_rate: maximal rate in near future (has to be evaluated externally)

rng: random number generator

API - output

returns the next event time

Generate a new event from an inhomogeneous poisson process with rate Lambda(t) under the assumption that rate(dt) is monotonically decreasing. Based on (Ogata’s Modified Thinning Algorithm: Ogata, 1981, p.25, Algorithm 3) see also https://www.math.fsu.edu/~ychen/research/Thinning%20algorithm.pdf

Arguments

rate: rate(dt) has to be defined outside (e..g t-> rate(t+t0,args))

rng: random number generator

API - output

returns the next event time

MonteCarloX.next_event — Function

next_event([rng::AbstractRNG,] rates::AbstractWeights)::Int

Select a single random index in 1:length(rates) with probability proportional to the entry in rates.

Remarks

This is on average twice as fast as StatsBase.sampling because it can iterate from either beginning or end of rates

next_event([rng::AbstractRNG,] event_handler::AbstractEventHandlerRate{T})::T where T

Select a single random event with a given probability managed by event_handler.

The event_handler also manages the case that no valid events are left (e.g. when all rates are equal to zero). This becomes relevant when using advance! to advance for some time.

See also: advance!

Generate a new event id from a collection of inhomogeneous poisson processes with rates Lambda(t).

Arguments

rates: rates(dt); Float -> [Float]

max_rate: maximal rate in near future (has to be evaluated externally)

rng: random number generator

API - output

returns the next event id

MonteCarloX.advance! — Function

advance!([rng::AbstractRNG], alg::KineticMonteCarlo, event_handler::AbstractEventHandlerRate, update!::Function, total_time::T)::T where {T<:Real}

Draw events from event_handler and update event_handler with update! until total_time has passed. Return time of last event.

advance!([rng::AbstractRNG], alg::KineticMonteCarlo, rates::Union{AbstractWeights,AbstractVector}, update!::Function, total_time::T)::T where {T<:Real}

Draw events from event_handler and update event_handler with update! until total_time has passed. Return time of last event.

Inhomogeneous Poisson Process